make upstream_failed acceptable for dependency check.
This is a temporary fix to make SG run successfully. Because there are
datasets failed due to its update_source.py failed but this shouldn't
block the SG build. Also, now it's possible to define allowed and not
allowed states, so next time we can just modify the DAGs to change
these values.