prevent DAG dependency check fail if etl script failed
perviously etl/recipe DAG will wait for the `validate` task for each
dependencies, but this will stop the DAG if the dependency DAG is also
etl/recipe and its etl script failed. Changing the task to be checked
to `cleanup` will make the DAG continue to run without problem because
the dependency DAG is still valid after clean-up.